chore(refactor): Split `oneil_output` errors into submodules
